+++ resources/views/layouts/app.blade.php
@@ -15,6 +15,8 @@
         ['route' => 'integrations.index', 'icon' => 'bolt', 'label' => __('Integrations'), 'match' => 'integrations*'],
         ['route' => 'developers.index', 'icon' => 'code', 'label' => __('Developers'), 'match' => 'developers*'],
         ['route' => 'tips.index', 'icon' => 'gift', 'label' => __('Tips'), 'match' => 'tips*'],
+        ['route' => 'credits.index', 'icon' => 'sparkles', 'label' => __('Credits'), 'match' => 'credits*'],
+        ['route' => 'activity.index', 'icon' => 'clock', 'label' => __('Activity'), 'match' => 'activity*'],
         ['route' => 'affiliate.index', 'icon' => 'gift', 'label' => __('Affiliate'), 'match' => 'affiliate*'],
         ['route' => 'billing.plans', 'icon' => 'card', 'label' => __('Billing'), 'match' => 'billing*'],
         ['route' => 'account.index', 'icon' => 'settings', 'label' => __('Account'), 'match' => 'account*'],